made segment select tool do resize & pencil as well, using m_dispatchTool
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@2312 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ gui/notationview.cpp
@@ -1930,7 +1930,7 @@ NotationView::slotCurrentStaffDown()
 {
     if (m_staffs.size() < 2) return;
     m_staffs[m_currentStaff]->setCurrent(false);
-    if (++m_currentStaff >= m_staffs.size()) m_currentStaff = 0;
+    if (++m_currentStaff >= (int)m_staffs.size()) m_currentStaff = 0;
     m_staffs[m_currentStaff]->setCurrent(true);
     slotSetInsertCursorPosition(m_insertionTime);
 }
